Fix testCommandScript test
