+++ src/test/java/io/antmedia/test/db/DBStoresUnitTest.java
@@ -7,6 +7,7 @@ import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
+import static org.mockito.Mockito.times;
 
 import java.io.File;
 import java.io.IOException;
@@ -2686,7 +2687,7 @@ public class DBStoresUnitTest {
 		String serverIP = "localhost";
 		
 		MongoStore dataStoreSpy = Mockito.spy(new MongoStore(serverIP, "", "", dbName));
-		assertTrue(dataStoreSpy.serverUpdated);
+		assertTrue(dataStoreSpy.isDbUpdated());
 	}
